--- a/lein-cljfmt/src/leiningen/cljfmt.clj
+++ b/lein-cljfmt/src/leiningen/cljfmt.clj
@@ -81,9 +81,12 @@
    (apply fix project (format-paths project)))
   ([project path & paths]
    (let [files (mapcat (partial find-files project) (cons path paths))]
-     (doseq [f files :when (not (valid-format? project f))]
+     (doseq [f     files
+             :let  [original (slurp f)
+                    revised  (reformat-string project original)]
+             :when (not= original revised)]
        (main/info "Reformatting" (project-path project f))
-       (spit f (reformat-string project (slurp f)))))))
+       (spit f revised)))))
 
 (defn cljfmt
   "Format Clojure source files"