This patch fixes #55, by inlining the `valid-format?` check and tweaking diff generation so that files are never read or formatted twice. This gives the expected 50% speedup.

before:
> lein cljfmt fix src/clj  772.30s user 1.78s system 101% cpu 12:41.43 total

after:
> lein cljfmt fix src/clj  396.35s user 1.40s system 103% cpu 6:23.34 total

merges cleanly with #54