@weavejester 2016-02-24
Thanks for the patch! Looking at your commits, it seems like 79307f2 has no purpose without 12052b7, so I think those two should be squashed together.

So perhaps we start with 4df8fc6, but let's change the commit message to simply:

    Don't reformat files twice in fix

The "how" is obvious from the diff, I think, so there's no need to explain it in the commit message.

Then we combine 79307f2 and 12052b7 into one, perhaps with the message:

    Don't reformat files twice in check

Again, I think the diff explains the "how".

You also should remove the `valid-format?` function, since it's no longer necessary after 12052b7.